Calculates industry energy shares for IEA data

Calculates both for input (from the U matrix) and for outputs (from the V matrix). All incoming PSUT matrix columns in psut_df are dropped. The output is two coumns of column vectors.

calc_iea_industry_shares(
  psut_df,
  trim_cols = TRUE,
  R = Recca::psut_cols$R,
  U = Recca::psut_cols$U,
  U_feed = Recca::psut_cols$U_feed,
  U_eiou = Recca::psut_cols$U_eiou,
  r_eiou = Recca::psut_cols$r_eiou,
  V = Recca::psut_cols$V,
  Y = Recca::psut_cols$Y,
  S_units = Recca::psut_cols$S_units,
  Y_fu_details = Recca::psut_cols$Y_fu_details,
  U_eiou_fu_details = Recca::psut_cols$U_eiou_fu_details,
  input_fractions_column = "U_fractions",
  output_fractions_column = "V_fractions"
)

Arguments

psut_df

A data frame of PSUT matrices

trim_cols

A boolean that tells whether to eliminate matrix columns on output. Default is TRUE.

R, U_feed, U_eiou, r_eiou, Y, S_units, Y_fu_details, U_eiou_fu_details

Columns deleted on output.

U

The name of the U matrix in psut_df. This column is deleted on output. Default is "U".

V

The name of the V matrix in psut_df. This column is deleted on output. Default is "V".

input_fractions_column

The name of the output column that contains vectors of input fractions.

output_fractions_column

The name of the output column that contains vectors of output fractions.

Value

A version of psut_df with input and output industry shares added as column vectors at the right side of the data frame.