Load an aggregation map

An aggregation file is an Excel file that contains an aggregation_tab with many_colname and few_colname, by default "Many" and "Few". This function loads the aggregation table and converts to an aggregation map using matsbyname::agg_table_to_agg_map().

load_agg_map(
  aggregation_file,
  aggregation_tab = tools::file_path_sans_ext(basename(aggregation_file)),
  country_colname = "Country",
  year_colname = "Year",
  many_colname = "Many",
  few_colname = "Few"
)

Arguments

aggregation_file

The path to the industry aggregations file.

aggregation_tab

The tab in industry_aggregations_file that contains an industry aggregation table. Default is aggregation_file |> basename() |> tools::file_path_sans_ext()

country_colname

The name of the country column in aggregation_tab. Default is "Country".

year_colname

The name of the year column in aggregation_tab. Default is "Year".

many_colname

The name of the many column on the industry_aggregations_tab. Default is "Many".

few_colname

The name of the few column on the industry_aggregations_tab. Default is "Few".

Value

An aggregation map.