Targeted product and industry aggregation according to an aggregation map

matsbyname::aggregate_byname() performs aggregation of rows and columns but is unaware of countries and years. This function performs targeted aggregation for some or all countries and years by including a country_colname and a year_colname in the aggregation_map. In addition, the Country and Year columns may contain optional "All" strings to indicate all countries or years should be aggregated as specified.

targeted_aggregation(
  psut_df,
  aggregation_map,
  country_colname = "Country",
  year_colname = "Year",
  agg_map_colname = "agg_map",
  margin = c("Industry", "Product"),
  matcols = c("R", "U", "U_feed", "U_EIOU", "r_EIOU", "V", "Y", "S_units"),
  matnames_colname = "matnames",
  matvals_colname = "matvals"
)

Arguments

psut_df

A data frame of PSUT matrices with country_colname, year_colname, and columns of PSUT matrices (matcols).

aggregation_map

An aggregation map with country_colname, year_colname, and agg_map_colname in which agg_map_colname contains aggregation maps to be applied for the country and year specified.

country_colname, year_colname, agg_map_colname

Names of columns in psut_df.

margin

The row or column types to be aggregated. Default is c("Product", "Industry").

matcols

A list of matrix column names.

matnames_colname, matvals_colname

Columns names used internally.

Value

A version of psut_df with aggregations according to aggregation_map and rowcoltype.